Social-Tools/Instagram/requirements.txt
instagrapi==1.17.10
python-dotenv==0.20.0
